+++ backend/ruoyi-admin-wms/src/main/java/com/ruoyi/wms/domain/bo/ItemSkuBo.java
@@ -83,7 +83,7 @@ public class ItemSkuBo extends BaseEntity {
     /**
      * 最低库存
      */
-    private BigDecimal minStock;
+    private Integer minStock;
 
     /**
      * 商品名称
