+++ backend/ruoyi-admin-wms/src/main/java/com/ruoyi/wms/domain/bo/MovementOrderBo.java
@@ -66,7 +66,7 @@ public class MovementOrderBo extends BaseEntity {
     /**
      * 总数量
      */
-    private BigDecimal totalQuantity;
+    private Integer totalQuantity;
 
     /**
      * 备注
