+++ backend/ruoyi-admin-wms/src/main/java/com/ruoyi/wms/domain/bo/ShipmentOrderBo.java
@@ -60,7 +60,7 @@ public class ShipmentOrderBo extends BaseEntity {
     /**
      * 出库数量
      */
-    private BigDecimal totalQuantity;
+    private Integer totalQuantity;
 
     /**
      * 出库单状态
