+++ backend/ruoyi-admin-wms/src/main/java/com/ruoyi/wms/domain/entity/ReceiptOrder.java
@@ -50,7 +50,7 @@ public class ReceiptOrder extends BaseEntity {
     /**
      * 商品总数
      */
-    private BigDecimal totalQuantity;
+    private Integer totalQuantity;
     /**
      * 订单金额
      */
